Move write-to-file
