+++ projects/optic/src/utils/write-to-file.test.ts
@@ -1,5 +1,5 @@
 import { test, expect, describe, beforeEach, jest } from '@jest/globals';
-import { addExtensionsToYaml } from '../write-to-file';
+import { addExtensionsToYaml } from './write-to-file';
 const yaml = `
 openapi: 3.0.1
 info:
